3d_renderer/requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw
numpy
pyrr
pillow
screeninfo
posix_ipc

apt
glfw error
libglfw3
libglfw3-dev

numpy error
libatlas-base-dev


GLSL 3.30 Is Not Supported
export MESA_GL_VERSION_OVERRIDE=3.3
